Fix: 로그아웃 함수를 헤더바가 아니라 각 페이지로 이전 (#77)
